[IMP] move `get_max_workers` function to `misc.py`
Such function doesn't belong to `pg.py`.
