+++ src/util/spreadsheet/misc.py
@@ -8,8 +8,8 @@ from functools import partial
 from odoo import sql_db
 
 from .. import json
-from ..misc import log_progress, make_pickleable_callback
-from ..pg import SQLStr, format_query, get_max_workers
+from ..misc import get_max_workers, log_progress, make_pickleable_callback
+from ..pg import SQLStr, format_query
 
 MEMORY_CAP = 2 * 10**8  # 200MB
 COUNT_CAP = 1000
